# feat: implement faster `get_device_type`