… test/cuda_tests.jl
… test/metal_tests.jl
… test/oneapi_tests.jl