```julia
julia> using LuxCUDA, LuxDeviceUtils, BenchmarkTools

julia> x = rand(10, 10) |> LuxCUDADevice();

julia> get_device_type()^C

julia> @code_typed get_device_type((x, x, x))
CodeInfo(
1 ─     return LuxCUDADevice
) => Type{LuxCUDADevice}

julia> @code_typed get_device((x, x, x))
CodeInfo(
1 ─      nothing::Nothing
│   %2 = invoke LuxDeviceUtils._get_device(x::Tuple{CuArray{Float32, 2, CUDA.DeviceMemory}, CuArray{Float32, 2, CUDA.DeviceMemory}, CuArray{Float32, 2, CUDA.DeviceMemory}})::LuxCUDADevice{CuDevice}
└──      return %2
) => LuxCUDADevice{CuDevice}

julia> @benchmark get_device_type(($x, $x, ($x, $x)))
BenchmarkTools.Trial: 10000 samples with 1000 evaluations.
 Range (min … max):  1.396 ns … 39.530 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     2.444 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   2.528 ns ±  0.640 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

                              ▃█ ▅         ▁ ▅                
  ▃▁▃▁▁▁▁▁▁▁▁▁▂▁▅▁▄▁▁▁▁▁▁▁▂▁▇▁██▁█▁▁▁▁▁▁▁▂▁█▁█▁▆▁▁▁▁▁▁▁▂▁▃▁▂ ▃
  1.4 ns         Histogram: frequency by time        3.35 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> @benchmark get_device(($x, $x, ($x, $x)))
BenchmarkTools.Trial: 10000 samples with 303 evaluations.
 Range (min … max):  274.756 ns …  1.738 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     300.802 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   310.144 ns ± 41.508 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

   ▆▇▇▆▇██▇▇▆▅▄▃▃▂▂▂▂▁▁▁  ▁                                    ▃
  █████████████████████████████▇██▇█▇▇███▇▇▆▆▆▆▆▆▆▅▆▆▄▆▆▇▆▇▇▇▇ █
  275 ns        Histogram: log(frequency) by time       486 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```